requirements.txt
google-cloud-aiplatform[reasoningengine,langchain]

google-genai
cloudpickle
requests
pillow
langchain-community
langchain-google-vertexai
google-cloud-storage<3.0.0
rich
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-logging
pandas
numpy
click
langchain-core
langgraph
google-auth
google-generativeai
google-genai-texttospeech

cloud-sql-python-connector
sqlalchemy
db-dtypes
google-cloud-texttospeech
google-cloud-speech
fastapi
notion-client
playwright
google-cloud-discoveryengine
watchdog